+++ tests/net/socket/can/prj.conf
@@ -1,0 +1,2 @@
+CONFIG_ZTEST=y
+CONFIG_ZTEST_NEW_API=y
