+++ tests/net/socket/can/testcase.yaml
@@ -1,0 +1,6 @@
+tests:
+  net.socket.can:
+    integration_platforms:
+      - native_posix
+      - native_posix_64
+    tags: net can
